hw: input : make the virtio input device configurable
Split out the virtio input device from the hw/virtio/virtio-pci.c
and make those virtio input devices configurable.
