+++ default-configs/pci.mak
@@ -44,3 +44,4 @@ CONFIG_IVSHMEM_DEVICE=$(CONFIG_IVSHMEM)
 CONFIG_ROCKER=y
 CONFIG_PCI_BRIDGE=y
 CONFIG_VFIO_VGA=y
+CONFIG_VIRTIO_INPUT_PCI=y
