+++ default-configs/x86_64_virt-softmmu.mak
@@ -12,7 +12,6 @@ CONFIG_VIRTIO_9P=y
 CONFIG_VIRTIO_BALLOON=y
 CONFIG_VIRTIO_BLK=y
 CONFIG_VIRTIO_CRYPTO=y
-CONFIG_VIRTIO_INPUT=y
 CONFIG_VIRTIO_NET=y
 CONFIG_VIRTIO_RNG=y
 CONFIG_VIRTIO_SCSI=y
